video: Extract software decoders (or glue) into their own modules, add screenvideo feature
