+++ core/Cargo.toml
@@ -58,6 +58,7 @@ approx = "0.5.0"
 default = ["minimp3", "serde"]
 h263 = ["h263-rs", "h263-rs-yuv"]
 vp6 = ["nihav_core", "nihav_codec_support", "nihav_duck", "h263-rs-yuv"]
+screenvideo = []
 lzma = ["lzma-rs", "swf/lzma"]
 wasm-bindgen = [ "instant/wasm-bindgen" ]
 avm_debug = []
