+++ core/src/backend/video/software/decoders/mod.rs
@@ -1,0 +1,8 @@
+#[cfg(feature = "h263")]
+pub mod h263;
+
+#[cfg(feature = "vp6")]
+pub mod vp6;
+
+#[cfg(feature = "screenvideo")]
+pub mod screen;
