+++ core/src/backend/video/software/mod.rs
@@ -1,0 +1,149 @@
+//! Pure software video decoding backend.
+
+use crate::backend::render::{BitmapHandle, BitmapInfo, RenderBackend};
+use crate::backend::video::{
+    DecodedFrame, EncodedFrame, Error, FrameDependency, VideoBackend, VideoStreamHandle,
+};
+use generational_arena::Arena;
+use swf::{VideoCodec, VideoDeblocking};
+
+mod decoders;
+
+#[cfg(feature = "h263")]
+use self::decoders::h263;
+
+#[cfg(feature = "vp6")]
+use self::decoders::vp6;
+
+#[cfg(feature = "screenvideo")]
+use self::decoders::screen;
+
+/// Software video backend that proxies to CPU-only codec implementations that
+/// ship with Ruffle.
+pub struct SoftwareVideoBackend {
+    streams: Arena<VideoStream>,
+}
+
+impl Default for SoftwareVideoBackend {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl SoftwareVideoBackend {
+    pub fn new() -> Self {
+        Self {
+            streams: Arena::new(),
+        }
+    }
+}
+
+impl VideoBackend for SoftwareVideoBackend {
+    #[allow(unreachable_code, unused_variables)]
+    fn register_video_stream(
+        &mut self,
+        _num_frames: u32,
+        size: (u16, u16),
+        codec: VideoCodec,
+        _filter: VideoDeblocking,
+    ) -> Result<VideoStreamHandle, Error> {
+        let decoder: Box<dyn VideoDecoder> = match codec {
+            #[cfg(feature = "h263")]
+            VideoCodec::H263 => Box::new(h263::H263Decoder::new()),
+            #[cfg(feature = "vp6")]
+            VideoCodec::Vp6 => Box::new(vp6::Vp6Decoder::new(false, size)),
+            #[cfg(feature = "vp6")]
+            VideoCodec::Vp6WithAlpha => Box::new(vp6::Vp6Decoder::new(true, size)),
+            #[cfg(feature = "screenvideo")]
+            VideoCodec::ScreenVideo => Box::new(screen::ScreenVideoDecoder::new()),
+            _ => return Err(format!("Unsupported video codec type {:?}", codec).into()),
+        };
+        let stream = VideoStream::new(decoder);
+        let stream_handle = self.streams.insert(stream);
+        Ok(stream_handle)
+    }
+
+    fn preload_video_stream_frame(
+        &mut self,
+        stream: VideoStreamHandle,
+        encoded_frame: EncodedFrame<'_>,
+    ) -> Result<FrameDependency, Error> {
+        let stream = self
+            .streams
+            .get_mut(stream)
+            .ok_or("Unregistered video stream")?;
+
+        stream.decoder.preload_frame(encoded_frame)
+    }
+
+    fn decode_video_stream_frame(
+        &mut self,
+        stream: VideoStreamHandle,
+        encoded_frame: EncodedFrame<'_>,
+        renderer: &mut dyn RenderBackend,
+    ) -> Result<BitmapInfo, Error> {
+        let stream = self
+            .streams
+            .get_mut(stream)
+            .ok_or("Unregistered video stream")?;
+
+        let frame = stream.decoder.decode_frame(encoded_frame)?;
+        let handle = if let Some(bitmap) = stream.bitmap {
+            renderer.update_texture(bitmap, frame.width.into(), frame.height.into(), frame.rgba)?
+        } else {
+            renderer.register_bitmap_raw(frame.width.into(), frame.height.into(), frame.rgba)?
+        };
+        stream.bitmap = Some(handle);
+
+        Ok(BitmapInfo {
+            handle,
+            width: frame.width,
+            height: frame.height,
+        })
+    }
+}
+
+/// A single preloaded video stream.
+struct VideoStream {
+    bitmap: Option<BitmapHandle>,
+    decoder: Box<dyn VideoDecoder>,
+}
+
+impl VideoStream {
+    fn new(decoder: Box<dyn VideoDecoder>) -> Self {
+        Self {
+            decoder,
+            bitmap: None,
+        }
+    }
+}
+
+/// Trait for video decoders.
+/// This should be implemented for each video codec.
+trait VideoDecoder {
+    /// Preload a frame.
+    ///
+    /// No decoding is intended to happen at this point in time. Instead, the
+    /// video data should be inspected to determine inter-frame dependencies
+    /// between this and any previous frames in the stream.
+    ///
+    /// Frames should be preloaded in the order that they are recieved.
+    ///
+    /// Any dependencies listed here are inherent to the video bitstream. The
+    /// containing video stream is also permitted to introduce additional
+    /// interframe dependencies.
+    fn preload_frame(&mut self, encoded_frame: EncodedFrame<'_>) -> Result<FrameDependency, Error>;
+
+    /// Decode a frame of a given video stream.
+    ///
+    /// This function is provided the external index of the frame, the codec
+    /// used to decode the data, and what codec to decode it with. The codec
+    /// provided here must match the one used to register the video stream.
+    ///
+    /// Frames may be decoded in any order that does not violate the frame
+    /// dependencies declared by the output of `preload_video_stream_frame`.
+    ///
+    /// The decoded frame should be returned. An `Error` can be returned if
+    /// a drawable bitmap can not be produced.
+    fn decode_frame(&mut self, encoded_frame: EncodedFrame<'_>) -> Result<DecodedFrame, Error>;
+}
