+++ desktop/Cargo.toml
@@ -29,13 +29,14 @@ winapi = "0.3.9"
 embed-resource = "1"
 
 [features]
-default = ["h263", "vp6"]
+default = ["h263", "vp6", "screenvideo"]
 
 # core features
 avm_debug = ["ruffle_core/avm_debug"]
 h263 = ["ruffle_core/h263"]
 lzma = ["ruffle_core/lzma"]
 vp6 = ["ruffle_core/vp6"]
+screenvideo =  ["ruffle_core/screenvideo"]
 
 # wgpu features
 render_debug_labels = ["ruffle_render_wgpu/render_debug_labels"]
